requirements.txt
altgraph==0.17.3
attrs==22.2.0
Babel==2.11.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
contextlib2==21.6.0
cssselect==1.2.0
cssutils==2.6.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
execnet==1.9.0
future==0.18.2
greenlet==2.0.1
idna==3.4
importlib-metadata==5.0.0
iniconfig==2.0.0
jaraco.classes==3.2.3
keyring==23.11.0
lxml==4.9.1
mock==5.0.1
more-itertools==9.0.0
mysql-connector-python==8.0.31
numpy==1.24.1
openpyxl==3.0.10
packaging==23.0
pandas==1.5.2
path==16.6.0
path.py==12.5.0
pdfkit==1.0.0
pefile==2022.5.30
Pillow==9.3.0
pluggy==1.0.0
premailer==3.10.0
protobuf==3.20.1
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pytest==7.2.1
python-dateutil==2.8.2
pytz==2022.6
pywin32-ctypes==0.2.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.46
termcolor==2.2.0
tkcalendar==1.6.1
tomli==2.0.1
ttkwidgets==0.12.1
urllib3==1.26.12
yagmail==0.15.293
zipp==3.10.0
